=== Refactor: impliments LoginForm component ===
